Move the code that runs in the page context to main-script.js.
Previously, Chrome required running in the main world because injecting
arbitrary script code from an extension is not allowed, but injecting a script
defined in the manifest is allowed. Firefox only supports running in the
isolated world, but did allow injecting the script.

The new setup works consistently between browsers: the extension runs in the
isolated world, and adds a <script src="main-script.js"> tag to the document
after it is loaded, and that script inspects the _preloads variable to extract
the post id, which is sent to the extension via the acx-page-load event.