Update comment API.
Removes the processHeader API function in favor of just using
processComment and passing in the ExtCommentComponent instead of the
data and comment element separately.

This simplifies the API and also allows options to access and modify
the comment component state. Since the comment data and element are
stored in the component, options can still access these.