Move comment time creation to option API.
Since there will be a few time related options, it makes sense to move
the logic out of the comment components. Right now, timeFormatOption is
a bit of a dummy option that doesn't store a value, but as more date
related options are added, it can use those values to correctly create
the timestamps.