Extract reverseDomChildren() function, and optimize the implementation somewhat.
It seems this is slightly faster in practice than putting childNodes in an array
and reversing all at once, because browsers keep children in a linked list, so
you can reverse the list in linear time.

Unfortunately switching the comment ordering doesn't become much faster. While
updating the dom takes < 100 ms after this change, the browser still needs
around 1 second to update the view, probably because it needs to recompute the
layout of all nodes after the change.